+++ src/library/PostgREST/Logger.hs
@@ -25,7 +25,6 @@ import Data.Text qualified as T
 import Data.Text.Encoding qualified as T
 
 import PostgREST.Config (LogLevel (..), Verbosity (..))
-import PostgREST.Config.DeprecatedJSPath (dumpDeprecatedJSPath)
 import PostgREST.Config.PgVersion (pgvName)
 import PostgREST.Debounce (makeDebouncer)
 import PostgREST.Logger.Apache (apacheFormat)
@@ -204,10 +203,6 @@ observationMessages = \case
   LegacyTargetNameWarningObs (warningMsg, warningHints) requestMethod requestTarget ->
     [ "WARNING: " <> warningMsg
     , "Update filters, orders or limits that use " <> warningHints <> " in " <> "`" <> T.decodeUtf8 (requestMethod <> " " <> requestTarget) <> "`"
-    ]
-  DeprecatedJSPathSyntaxObs djsp ->
-    [ "WARNING: The config `jwt-role-claim-key=" <> dumpDeprecatedJSPath djsp <> "` is using the deprecated JSPath syntax."
-    , "Update `jwt-role-claim-key=" <> dumpDeprecatedJSPath djsp <> "` to the new JSONPath syntax as support for JSPath will be removed in a future release. See the migration guide on how to update: https://github.com/PostgREST/postgrest/blob/main/CHANGELOG.md#migration-to-v16"
     ]
   ConfigReadErrorObs usageErr ->
     pure $ "Failed to query database settings for the config parameters." <> jsonMessage usageErr